gsoc-gemini-integration/requirements.txt
langchain
llama-index
composio
crewai
google-generativeai
openai
google-generativeai
google-generativeai
pytest
